feat(test): Implement execute and query contract
