--- test/chains/archway/archway_chain.go
@@ -1,18 +1,0 @@
-package archway
-
-import (
-	"fmt"
-	"github.com/icon-project/ibc-integration/test/chains"
-)
-
-func NewArchwayChain(environment string, chainConfig chains.ChainConfig) (chains.Chain, error) {
-	switch environment {
-	case "local", "localnet":
-
-	case "testnet":
-	default:
-		return nil, fmt.Errorf("unknown environment: %s", environment)
-	}
-
-	return nil, nil
-}
