+++ test/chains/types.go
@@ -4,3 +4,19 @@ type ContractKey struct {
 	ContractAddress string
 }
 type LastBlock struct{}
+
+type ContractName struct {
+	ContractName string
+}
+
+type InitMessage struct {
+	InitMsg string
+}
+
+type Param struct {
+	Data string
+}
+
+type Query struct {
+	Query string
+}
